+++ spec/logger_spec.rb
@@ -28,6 +28,66 @@ describe BufferingLogger::Logger do
         end
         expect(dev_contents).to eq message
       end
+
+      context 'with multithreaded logging' do
+        before do
+          # format the log with just the message plus a newline
+          logger.formatter = ->(_, _, _, msg) { "#{msg}\n" }
+        end
+
+        it 'uses separate buffers for each thread' do
+          thread1_queue = Queue.new
+          thread2_queue = Queue.new
+
+          thread1 = Thread.new do
+            logger.buffered do
+              thread1_queue << 'started'
+              thread2_queue.shift == 'started' || raise
+
+              logger.info 'thread1 message 1'
+              thread1_queue << 'sent 1'
+
+              thread2_queue.shift == 'sent 1' || raise
+              logger.info 'thread1 message 2'
+              thread1_queue << 'sent 2'
+
+              thread2_queue.shift == 'sent 2' || raise
+            end
+
+            thread1_queue << 'finished'
+          end
+
+          thread2 = Thread.new do
+            logger.buffered do
+              thread2_queue << 'started'
+              thread1_queue.shift == 'started' || raise
+
+              thread1_queue.shift == 'sent 1' || raise
+
+              logger.info 'thread2 message 1'
+              thread2_queue << 'sent 1'
+
+              thread1_queue.shift == 'sent 2' || raise
+              logger.info 'thread2 message 2'
+              thread2_queue << 'sent 2'
+
+              thread1_queue.shift == 'finished' || raise
+            end
+          end
+
+          thread1.join
+          thread2.join
+
+          expected_grouping = <<-LOG.gsub(/^\s*/, '')
+            thread1 message 1
+            thread1 message 2
+            thread2 message 1
+            thread2 message 2
+          LOG
+
+          expect(dev_contents).to eq(expected_grouping)
+        end
+      end
     end
   end